:art: :tractor: Move fs.Repository implementation and add interface
