+++ pkg/repositories/fs/repository_fuzz_test.go
@@ -1,4 +1,4 @@
-package repositories
+package fs
 
 import (
 	"github.com/go-go-golems/glazed/pkg/cmds"
